RDM-5734 Change search results to have more detail
Changes the search results to have source repo, creators and issued date
as well as the title. It adds the ability to have source repo logos if
the document has one. It changes the sort options and moves the
pagination and sorting to be above just the search results. It also
increases the screen width used to 95% if it is above the largest
breakpoint for bootstrap to better use the screen real estate on larger
screens.