RDM-5841 Make list of search results scroll by itself
This change is to make sure that the map and filters always display so
that when hovering over a result you can see the bounding box
highlighted on the map. It also fixes some sizing issues between the
list of results and the map across all breakpoints. Increases the real
estate if the display is large and resizes some fonts to make the paging
and sort results control work better.