Clustering with facets (#14)
Changed clustering call on results page to make a POST call to Globus search only returning the results filtered by the map, search terms, and map bounds. The returned JSON is then parsed into records that get added to PruneCluster to be clustered into markers. Currently only doing a maximum of 1000 records to cluster on the page due to time constraints.